+++ Formula/iterion.rb
@@ -1,28 +1,28 @@
 class Iterion < Formula
   desc "Build, run and orchestrate agentic AI workflows, from readable .bot files"
   homepage "https://github.com/SocialGouv/iterion"
-  version "3.174.0"
+  version "3.175.1"
   license "MIT"
 
   on_macos do
     on_arm do
       url "https://github.com/SocialGouv/iterion/releases/download/v#{version}/iterion-darwin-arm64"
-      sha256 "22f66d1a3daa728e1b26d3a39c19e2773c281eccd98276ba44eb8be2ee6cf922"
+      sha256 "06603ecdab6cdcfac4ff2a4322283e0769e0738f7be04324b99727243d5ca459"
     end
     on_intel do
       url "https://github.com/SocialGouv/iterion/releases/download/v#{version}/iterion-darwin-amd64"
-      sha256 "6ee3bd0347104e4f4f400db17628173307f404e06ed267b99212e27a23cab601"
+      sha256 "ab3d824e402e84d0a1d417e036ce18543e3a69d36673b730372208287dd36261"
     end
   end
 
   on_linux do
     on_arm do
       url "https://github.com/SocialGouv/iterion/releases/download/v#{version}/iterion-linux-arm64"
-      sha256 "8d71c3309967a04b7dc0a6c6bca4ee5d79b61e4e262012946a1ccaa2bfae408f"
+      sha256 "2c2151a4ef3e35b7e9ed0864f3328e9cf79a03a13a5668c82afff635f45aac42"
     end
     on_intel do
       url "https://github.com/SocialGouv/iterion/releases/download/v#{version}/iterion-linux-amd64"
-      sha256 "a8564b53efc852f7186b3f7e01c96bd494918fdb487d9308238de0a40b6a87a8"
+      sha256 "50e1cc8406c0ec0eac5be19115ccd7ad23ae454e5099dcf8da0327ddd3724d36"
     end
   end